Fix for SOS-2455
Fixed SOS-2455: PWC: XiO adapter stuck in API Error state and CPX had to
be restarted. The API status check was done before the monitoring
timestamp request was updated. Because of that, if API error lasted more
than pause monitoring timeout (3 min), device was considered as with
monitoring paused and new statistics was not retrieved from XiO.